Move from onHook to event subscribing.
By simplifiyng logic, errors can be avoided more easily.
However, this means this is run only when the display is on a host. I do not know how that would work with networking.
Or how that works with syncing to clients as I have not looked very well in how this mod works.